+++ src/schemathesis/models.py
@@ -938,7 +938,7 @@ class TestResult:
     method: str = attr.ib()  # pragma: no mutate
     path: str = attr.ib()  # pragma: no mutate
     verbose_name: str = attr.ib()  # pragma: no mutate
-    data_generation_method: DataGenerationMethod = attr.ib()  # pragma: no mutate
+    data_generation_method: List[DataGenerationMethod] = attr.ib()  # pragma: no mutate
     checks: List[Check] = attr.ib(factory=list)  # pragma: no mutate
     errors: List[Tuple[Exception, Optional[Case]]] = attr.ib(factory=list)  # pragma: no mutate
     interactions: List[Interaction] = attr.ib(factory=list)  # pragma: no mutate